Move: Move files to fit folder structure
