+++ app/src/main/java/com/example/sweatnote/components/writing/CheckboxWithText.kt
@@ -22,10 +22,9 @@ import com.example.sweatnote.ui.theme.pretendardFamily
 @Composable
 fun CheckboxWithText(text: String, isChecked: Boolean, onCheckedChange: (Boolean) -> Unit) {
     Surface(
-        shape = RoundedCornerShape(12.dp), // 테두리를 더욱 둥글게 만듭니다.
-        color = if (isChecked) DarkGray else Color(0xFFF5F5F5), // 체크박스가 체크되면 배경색을 진한 회색으로, 아니면 연한 회색으로 설정합니다.
-        contentColor = if (isChecked) White else Black, // 체크박스가 체크되면 글자색을 하얀색으로, 아니면 검은색으로 설정합니다.
-//        border = if (isChecked) null else BorderStroke(width = 1.dp, color = Color.Gray), // 체크박스가 체크되면 테두리를 없애고, 아니면 회색 테두리를 유지합니다.
+        shape = RoundedCornerShape(12.dp),
+        color = if (isChecked) DarkGray else Color(0xFFF5F5F5),
+        contentColor = if (isChecked) White else Black,
         modifier = Modifier
             .clickable { onCheckedChange(!isChecked) }
             .width(100.dp)
